+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project does NOT adhere to [Semantic Versioning](https://semver.org/spe
 * grassfish.py: Add `insecure=False, no_proxy=False, timeout=8` to `fetch_json()`
 * huawei.py: no more hardcoded `insecure` parameter in `get_creds()` and `get_data()`
 * icinga.py: no more hardcoded `insecure` parameter; add `insecure=False, no_proxy=False, timeout=3` to all functions
+* infomaniak.py: add `insecure=False, no_proxy=False, timeout=8` to all functions
 * librenms.py: Introduce caching of the result sets.
 
 
@@ -25,7 +26,7 @@ and this project does NOT adhere to [Semantic Versioning](https://semver.org/spe
 
 ### Added
 
-* qts.py, a library for interacting with the QNAP QTS API
+* qts.py: A library for interacting with the QNAP QTS API
 * endoflifedate.py (built automatically; version string only changes when products are added or deleted)
 * [Published on PyPI](https://pypi.org/project/linuxfabrik-lib/), you may now use `pip install linuxfabrik-lib`
 * tools/update-endoflifedate, which - like its name says - can be used to update endoflifedate.py